requirements.txt
line-bot-sdk
django
requests
selenium